+++ invokeai/frontend/web/src/app/store/middleware/listenerMiddleware/listeners/appStarted.ts
@@ -1,5 +1,8 @@
 import { createAction } from '@reduxjs/toolkit';
-import { isInitializedChanged } from 'features/gallery/store/gallerySlice';
+import {
+  INITIAL_IMAGE_LIMIT,
+  isLoadingChanged,
+} from 'features/gallery/store/gallerySlice';
 import { receivedPageOfImages } from 'services/api/thunks/image';
 import { startAppListening } from '..';
 
@@ -20,22 +23,21 @@ export const addAppStartedListener = () => {
           categories: ['general'],
           is_intermediate: false,
           offset: 0,
-          limit: 100,
+          limit: INITIAL_IMAGE_LIMIT,
         })
       );
 
       // fill up the assets tab with images
-      dispatch(
+      await dispatch(
         receivedPageOfImages({
           categories: ['control', 'mask', 'user', 'other'],
           is_intermediate: false,
           offset: 0,
-          limit: 100,
+          limit: INITIAL_IMAGE_LIMIT,
         })
       );
 
-      // tell the gallery it has made its initial fetches
-      dispatch(isInitializedChanged(true));
+      dispatch(isLoadingChanged(false));
     },
   });
 };
